# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestão de ordens de serviço, clientes, veículos, serviços e estoque de peças.",
        "title": "Oficina Mecânica API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "TechChalleng",
            "email": "suporte@oficina.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Realiza login e retorna token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Autenticação de administrador",
                "parameters": [
                    {
                        "description": "Credenciais",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/clientes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clientes"
                ],
                "summary": "Listar clientes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.ClienteResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cadastra um novo cliente (requer JWT admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clientes"
                ],
                "summary": "Criar cliente",
                "parameters": [
                    {
                        "description": "Dados do cliente",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.CriarClienteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.ClienteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/clientes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clientes"
                ],
                "summary": "Buscar cliente por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.ClienteResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "clientes"
                ],
                "summary": "Excluir cliente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/pecas": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "pecas"
                ],
                "summary": "Listar peças",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.PecaResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pecas"
                ],
                "summary": "Criar peça/insumo",
                "parameters": [
                    {
                        "description": "Dados da peça",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.CriarPecaRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.PecaResponse"
                        }
                    }
                }
            }
        },
        "/pecas/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "pecas"
                ],
                "summary": "Buscar peça por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da peça",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.PecaResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "pecas"
                ],
                "summary": "Excluir peça",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da peça",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/servicos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "servicos"
                ],
                "summary": "Listar serviços",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.ServicoResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "servicos"
                ],
                "summary": "Criar serviço",
                "parameters": [
                    {
                        "description": "Dados do serviço",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.CriarServicoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.ServicoResponse"
                        }
                    }
                }
            }
        },
        "/servicos/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "servicos"
                ],
                "summary": "Buscar serviço por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do serviço",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.ServicoResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "servicos"
                ],
                "summary": "Excluir serviço",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do serviço",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/veiculos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "veiculos"
                ],
                "summary": "Listar veículos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.VeiculoResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "veiculos"
                ],
                "summary": "Criar veículo",
                "parameters": [
                    {
                        "description": "Dados do veículo",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.CriarVeiculoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.VeiculoResponse"
                        }
                    }
                }
            }
        },
        "/veiculos/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "veiculos"
                ],
                "summary": "Buscar veículo por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do veículo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.VeiculoResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "veiculos"
                ],
                "summary": "Excluir veículo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do veículo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_techchalleng_oficina_internal_application_dto.ClienteResponse": {
            "type": "object",
            "properties": {
                "cnpj": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "criado_em": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                }
            }
        },
        "github_com_techchalleng_oficina_internal_application_dto.CriarClienteRequest": {
            "type": "object",
            "required": [
                "email",
                "nome",
                "telefone"
            ],
            "properties": {
                "cnpj": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                }
            }
        },
...
        "github_com_techchalleng_oficina_internal_application_dto.CriarPecaRequest": {
            "type": "object",
            "required": [
                "codigo",
                "nome",
                "quantidade_estoque",
                "valor_unitario"
            ],
            "properties": {
                "codigo": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "quantidade_estoque": {
                    "type": "integer",
                    "minimum": 0
                },
                "valor_unitario": {
                    "type": "number"
                }
            }
        },
        "github_com_techchalleng_oficina_internal_application_dto.CriarServicoRequest": {
            "type": "object",
            "required": [
                "nome",
                "valor_mao_obra"
            ],
            "properties": {
                "descricao": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "valor_mao_obra": {
                    "type": "number"
                }
            }
        },
...
        "github_com_techchalleng_oficina_internal_application_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_techchalleng_oficina_internal_application_dto.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_techchalleng_oficina_internal_application_dto.LoginResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "github_com_techchalleng_oficina_internal_application_dto.PecaResponse": {
            "type": "object",
            "properties": {
                "ativo": {
                    "type": "boolean"
                },
                "codigo": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "quantidade_estoque": {
                    "type": "integer"
                },
                "valor_unitario": {
                    "type": "number"
                }
            }
        },
        "github_com_techchalleng_oficina_internal_application_dto.ServicoResponse": {
            "type": "object",
            "properties": {
                "ativo": {
                    "type": "boolean"
                },
                "descricao": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "valor_mao_obra": {
                    "type": "number"
                }
            }
        },
...
        "github_com_techchalleng_oficina_internal_application_dto.VeiculoResponse": {
            "type": "object",
            "properties": {
                "ano": {
                    "type": "integer"
                },
                "cliente_id": {
                    "type": "string"
                },
                "criado_em": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "marca": {
                    "type": "string"
                },
                "modelo": {
                    "type": "string"
                },
                "placa": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Insira o token JWT no formato: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}